Splitted and improved NavigationBinding options.
